broadway: Make the InputMessage type a uint32
That way we don't get any holes in the structs.

+++ gdk/broadway/broadway-protocol.h
@@ -39,7 +39,7 @@ typedef enum {
 } BroadwayOpType;
 
 typedef struct {
-  guint8 type;
+  guint32 type;
   guint32 serial;
   guint64 time;
 } BroadwayInputBaseMsg;
